==> 02_CellChat.Rmd <==
---
title: "CellChat"
author: "Qiuyun Yang"
date: "`r Sys.Date()`"
output: html_document
---
#Cellchat
mysubtypecolors<- c("#3b6291","#943c39","#779043","#624c7c","#388498","#bf7334","#ffa510","#f74d4d")
setwd("D:/R/TCF7L1/")

library(CellChat)
library(tidyverse)
library(ggalluvial)
library(Seurat)
library(data.table)
library(ggsci)

load("D:/R/TCF7L1/immune.combined.sct.named.RData")
sc.sp=SplitObject(immune.combined.sct,split.by = "orig.ident")
Control=immune.combined.sct[,sample(colnames(sc.sp[["Control"]]))]
mTBI=immune.combined.sct[,sample(colnames(sc.sp[["mTBI"]]))]
rm(immune.combined.sct)
rm(sc.sp)
cellchat.Control <- createCellChat(object =Control@assays$RNA@data, meta =Control@meta.data,  group.by ="celltype")
cellchat.mTBI <- createCellChat(object =mTBI@assays$RNA@data, meta =mTBI@meta.data,  group.by ="celltype")

dir.create("compare")
setwd("compare/")

cellchat=cellchat.Control
cellchat@DB  <- subsetDB(CellChatDB, search = "Secreted Signaling") # use Secreted Signaling
cellchat <- subsetData(cellchat)
future::plan("multiprocess", workers = 4)
cellchat <- identifyOverExpressedGenes(cellchat)
cellchat <- identifyOverExpressedInteractions(cellchat)
cellchat <- projectData(cellchat, PPI.mouse)
cellchat <- computeCommunProb(cellchat, raw.use = TRUE,population.size =T)
cellchat <- filterCommunication(cellchat, min.cells = 3)
cellchat <- computeCommunProbPathway(cellchat)
cellchat <- aggregateNet(cellchat)
cellchat <- netAnalysis_computeCentrality(cellchat, slot.name = "netP")
cc.Chontrol = cellchat
#################################
cellchat=cellchat.mTBI
cellchat@DB  <- subsetDB(CellChatDB, search = "Secreted Signaling") # use Secreted Signaling
cellchat <- subsetData(cellchat)
future::plan("multiprocess", workers = 4)
cellchat <- identifyOverExpressedGenes(cellchat)
cellchat <- identifyOverExpressedInteractions(cellchat)
cellchat <- projectData(cellchat, PPI.mouse)
cellchat <- computeCommunProb(cellchat, raw.use = TRUE,population.size =T)
cellchat <- filterCommunication(cellchat, min.cells = 3)
cellchat <- computeCommunProbPathway(cellchat)
cellchat <- aggregateNet(cellchat)
cellchat <- netAnalysis_computeCentrality(cellchat, slot.name = "netP")
cc.mTBI = cellchat
##############################################
cc.list=list(C=cc.Chontrol,mT=cc.mTBI)
cellchat=mergeCellChat(cc.list,cell.prefix = T,add.names = names(cc.list))
saveRDS(cellchat,file = "cellchat_comparisonAnalysis_control_mTBI.rds")

compareInteractions(cellchat,show.legend = F,group = c(1,3),measure = "count")
compareInteractions(cellchat,show.legend = F,group = c(1,3),measure = "weight")

netVisual_diffInteraction(cellchat,weight.scale = T)
netVisual_diffInteraction(cellchat,weight.scale = T,measure = "weight")

netVisual_heatmap(cellchat)
netVisual_heatmap(cellchat,measure = "weight")

rankNet(cellchat,mode = "comparison",stacked = T,do.stat = T)
rankNet(cellchat,mode = "comparison",stacked =F,do.stat = T)
